cheri: kernel: debug: Fabricate capabilities in the debugger
When debugging the kernel explicitly fabricate capabilities
to access kernel memory. We don't want the debugger to get
CHERI faults because of broken pointers in the debugee.

+++ include/linux/kgdb.h
@@ -73,7 +73,7 @@ enum kgdb_bpstate {
 };
 
 struct kgdb_bkpt {
-	unsigned long		bpt_addr;
+	uintptr_t		bpt_addr;
 	unsigned char		saved_instr[BREAK_INSTR_SIZE];
 	enum kgdb_bptype	type;
 	enum kgdb_bpstate	state;
